+++ Group-3_Warehouse-Management-System/src/main/java/com/example/dto/ProductItemDTO.java
@@ -11,10 +11,10 @@ public class ProductItemDTO {
     private Long productId;
     private String productName;
     private String serial;
-    private Long importPrice;
+    private Double importPrice;
     private String unit;
 
-    public ProductItemDTO(Long productId, String serial, Long importPrice, String unit) {
+    public ProductItemDTO(Long productId, String serial, Double importPrice, String unit) {
         this.productId = productId;
         this.serial = serial;
         this.importPrice = importPrice;
